test(service): add auth service unit tests with mocks
Test Coverage:
- Register success case
- Register with existing email
- Login success case
- Login with invalid credentials
- Login with wrong password
- Logout functionality

Mock Implementations:
- MockUserRepository for user data access
- MockRefreshTokenRepository for token storage
- Isolated testing without database dependency

Testing Strategy:
✅ Use mocks to isolate business logic
✅ Test success and error scenarios
✅ Verify repository method calls
✅ Assert correct error types returned
✅ Validate response structure

Benefits:
- Fast test execution (no DB)
- Reliable tests (no external dependencies)
- Easy to test edge cases
- Clear separation of concerns

All auth service tests passing

+++ internal/dto/common_dto.go
@@ -2,10 +2,10 @@ package dto
 
 // PaginationMeta represents pagination metadata
 type PaginationMeta struct {
-	Page       int `json:"page"`
-	PageSize   int `json:"page_size"`
-	TotalItems int `json:"total_items"`
-	TotalPages int `json:"total_pages"`
+	Page       int   `json:"page"`
+	PageSize   int   `json:"page_size"`
+	TotalItems int64 `json:"total_items"`
+	TotalPages int   `json:"total_pages"`
 }
 
 // ErrorResponse represents error response
